Added `RequiresAuthCard`
